requirements.txt
streamlit
python-dotenv
sqlite3
openai